requirements.txt
torch
torchvision
numpy
matplotlib
seaborn
scikit-learn
deap
tqdm
pillow
opencv-python
tkinterdnd2
